Agent_Frameworks/CrewAI/requirements.txt
# ── CrewAI core ───────────────────────────────────────────────────────────────
# CrewAI requires Python >=3.10,<3.14. Use python3 (3.10/3.11/3.12/3.13) to install.
# On this machine: python3 -m venv .venv && .venv\Scripts\activate && pip install -r requirements.txt
crewai>=0.100.0,<2.0.0
crewai-tools>=0.33.0,<2.0.0

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=1.13.3

# ── Configuration & environment ───────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Custom tool dependencies ──────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=2.1.0

# ── Terminal display ──────────────────────────────────────────────────────────
rich>=13.0.0
